requirements.txt
yfinance
python-telegram-bot
schedule
python-dotenv
flask
pytz
requests
telegram